+++ Retrospective/Core/CommonViews/CustomTextField.swift
@@ -6,17 +6,11 @@
 //
 
 import SwiftUI
-///TextField 커스텀 하실때 쓰시면됩니다. 아래는 사용 예시 입니다. 스트링 타입의 상태 변수 하나 만들어주시고 placeholder라는 파라미터에 원하시는 값 입력하시면 됩니다.
-//		@State private var nameInput: String = ""
-//
-//		var body: some View {
-//      CustomTextField(placeholder: "이름을 입력해주세요", text: $nameInput)
-//  }
 
 struct CustomTextField: View {
+
     var placeholder: String
     @Binding var text: String
-
 
     var body: some View {
         HStack{
